refactor: FireflyEndpointProvider
